Fixed bug with list keys and refactored Hexagon into separate file
